Update Block class

+++ src/config.hpp
@@ -31,7 +31,7 @@ constexpr float CROSSHAIR_X = 30.f / 1280.f, CROSSHAIR_Y = 32.f / 960.f, CROSSHA
 constexpr uint64_t DAYTIME = 600; // sec
 
 constexpr uint32_t SUB_TEX_WIDTH = 64, SUB_TEX_HEIGHT = 64, N_TILES = 5;
-constexpr uint8_t N_MIP_LEVEL = 7;
+constexpr int32_t N_MIP_LEVEL = 7;
 
 constexpr uint8_t
     FACE_LEFT   = 0, FACE_LEFT_BIT   = 1u << 0u,
