+++ src/ray.cpp
@@ -8,9 +8,9 @@ Block* Ray::cast_block(vec3 const& p0, vec3 const& dir, float max_distance) {
     array<int32_t, 3> step;
     array<float, 3>   next_distance;
 
-    auto compute_next_distance = [&](uint8_t i) { next_distance[i] = (static_cast<float>(curr[i] + (step[i] < 0 ? 0 : 1)) - p0[i]) / dir[i]; };
+    auto compute_next_distance = [&](int i) { next_distance[i] = (static_cast<float>(curr[i] + (step[i] < 0 ? 0 : 1)) - p0[i]) / dir[i]; };
 
-    for (uint8_t i = 0; i < 3; i++) {
+    for (int i = 0; i < 3; i++) {
         if (dir[i] != 0.f) {
             curr[i] = static_cast<int32_t>(floor(p0[i]));
             step[i] = dir[i] > 0.f ? 1 : -1;
@@ -21,7 +21,7 @@ Block* Ray::cast_block(vec3 const& p0, vec3 const& dir, float max_distance) {
     }
 
     for (;;) {
-        uint8_t i = 0;
+        int i = 0;
         if (next_distance[1] < next_distance[i])
             i = 1;
         if (next_distance[2] < next_distance[i])
